using float

+++ be/DatabaseSchemeApp.Domain/Articol.cs
@@ -6,7 +6,7 @@
         public int IdComanda { get; set; }
         public int IdStoc { get; set; }
         public int Cantitate { get; set; }
-        public decimal PretTotal { get; set; }
+        public float PretTotal { get; set; }
 
         public Comanda Comanda { get; set; }
         public Stoc Stoc { get; set; }
